refactor: split semantic_tokens.rs into module directory
Break the 1838-line monolith into focused submodules:

  mod.rs     (280 lines) — legend, types, public API, delta encoding
  kotlin.rs  (259 lines) — Phase 1 CST Kotlin declaration classification
  java.rs    (218 lines) — Phase 1 CST Java declaration classification
  helpers.rs (369 lines) — shared tree-sitter helpers, node predicates
  params.rs  (156 lines) — Phase 1b parameter use-site coloring
  resolve.rs (568 lines) — Phase 2 index-based reference resolution

Java-only modifier helpers moved into java.rs per review feedback.
All 761 tests pass, clippy clean. No functional changes.
